Fixed SDL Event Polling
You guys were not polling SDL for events.  This was causing the event
loop to constantly block.  The loop is now working properly but you guys
have to figure it out from here.  It tool a LOT of changes to get this
in here and to me it represents a fundamental misunderstanding by your
team about what system you are really using.  You are using SDL!
Everyone needs to know this and needs to structure their code
accordingly.

More importantly, I am very troubled to see that no changes have been
committed since yesterday.  You customer gave you a very real deadline
for today and right now it looks like you are not going to make it.
This reflects poorly on all of us on this team and me in the class and
will affect your grade for this sprint.  This is an all-hands-on-deck
moment yet I was only contacted by John and I do not see anyone else
even trying to get this working.  You guys need to wake up and take this
project more seriously!!  I'm here to help but if you aren't asking for
it and putting in your own effort things are not going to end well.