Add Animation and AnimationState Classes
Preparing to hook up new functionality to the animation component

My idea is to introduce animation states that are governed by the
controller. The controller will populate the animation state list (held
by the animation component) as it sees fit. Each animation state will
have an array of animations where the cells hold the directions which
the animation should face.

Next step: the animation component will be simplified to draw animations
based on which state is active. The controllers will gain functionality
to add animation states to the list.

Also added return statements to vInit and vUpdate in
ScriptableComponent.cpp as not having them was causing errors.